button added for receive docs and common code now used for both receive and transfer
